REST API into Interface and /api/ URL
