--- Source/Chronozoom.UI/Chronozoom.svc
@@ -1,1 +1,0 @@
-﻿<%@ ServiceHost Language="C#" Debug="true" Service="Chronozoom.UI.ChronozoomSVC" CodeBehind="Chronozoom.svc.cs" Factory="System.ServiceModel.Activation.WebScriptServiceHostFactory" %>
