+++ Source/Chronozoom.UI/Web.config
@@ -62,7 +62,7 @@
     <serviceHostingEnvironment aspNetCompatibilityEnabled="true" multipleSiteBindingsEnabled="true" />
     <services>
       <service name="Chronozoom.UI.ChronozoomSVC">
-        <endpoint address="" behaviorConfiguration="Chronozoom.UI.ChronozoomAspNetAjaxBehavior" binding="webHttpBinding" contract="Chronozoom.UI.ChronozoomSVC" />
+        <endpoint address="" behaviorConfiguration="Chronozoom.UI.ChronozoomAspNetAjaxBehavior" binding="webHttpBinding" contract="Chronozoom.UI.IChronozoomSVC" />
       </service>
     </services>
   </system.serviceModel>
