+++ Source/Chronozoom.UI/api/Chronozoom.svc
@@ -1,0 +1,1 @@
+﻿<%@ ServiceHost Language="C#" Debug="true" Service="Chronozoom.UI.ChronozoomSVC" CodeBehind="chronozoom.svc.cs" Factory="System.ServiceModel.Activation.WebScriptServiceHostFactory" %>
