+++ Source/Chronozoom.UI/scripts/common.ts
@@ -204,7 +204,7 @@ module CZ {
             } else {
                 switch (CZ.Settings.czDataSource) {
                     case 'db':
-                        return "/Chronozoom.svc/get";
+                        return "/api/get";
                     case 'relay':
                         return "ChronozoomRelay";
                     case 'dump':
